Neurons and Connctions are removeable now
If you remove a neuron, it will remove all connections this neuron is
part of, it will unregister itself from its network and it will disable
itself. A disabled neuron is not anymore able to calculate outputs and
it will return 0 as its current value.
Connections can also be removed, they will inform their input and output
neurons.